+++ src/weapon_draw.cpp
@@ -455,7 +455,7 @@ void draw_weapon(point const &pos, vector3d dir, float cradius, int cid, int wid
 				shader.add_uniform_float("min_alpha", 0.95*alpha);
 				shader.set_specular(0.9, 90.0);
 				float dz((ammo > 1) ? -0.025*radius*ammo : 0.0);
-				select_texture(sb_tex ? SAW_B_TEX : SAW_TEX);
+				select_texture(sb_tex ? (int)SAW_B_TEX : (int)SAW_TEX);
 
 				for (int w = 0; w < max(1, ammo); ++w) { // draw a blade for each ammo
 					draw_tquad(radius, radius, dz);
